"/update-status" - Require authentication
Note that there are two authentication mechanisms involved with this route,
but they do different things.

STATUS_SECRET is an internal value used by probot to generate a
STATUS_TOKEN.  The STATUS_TOKEN specifically grants access to update the
commit-status.  This ensures that Jenkins can only submit commit-statuses
for which we've recently given it a token.

Unfortunately, those tokens aren't secret.  A user inspecting the Jenkins UI
can discover the values of any *parameters* that were passed into the job.
To ensure that *only* Jenkins (and not some intrepid external user) can use
these tokens, we also setup HTTP basic auth -- the necessary password is
managed by Jenkins in a more sensitive manner.